requirements.txt
pymupdf
faiss-cpu
numpy
sentence-transformers
groq
requests
streamlit